pyproject.toml
[tool.poetry]
name = "useful-course-back"
version = "1.5.0"
description = ""
authors = ["DJWOMS <djwoms@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.6"
Django = "4.0.1"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.1"
drf-yasg = "^1.20.0"
Pillow = "^8.2.0"
django-mptt = "^0.13.4"
django-ckeditor = "^6.1.0"
django-cors-headers = "^3.7.0"
google-auth = "1.32.0"
httpx = "^0.21.1"
gunicorn = "^20.1.0"
PyJWT = "^2.1.0"
PyGithub = "^1.55"
django-ninja = "^0.16.1"
pydantic = {extras = ["email"], version = "^1.8.2"}
channels = "^3.0.4"
channels-redis = "^3.3.1"
djangochannelsrestframework = "^0.3.0"
daphne = "^3.0.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.21.0"}
uvicorn = {extras = ["standart"], version = "^0.16.0"}
redis = "^4.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
